+++ final-pjt-back/paradiso/settings.py
@@ -143,6 +143,6 @@ AUTH_USER_MODEL = "accounts.User"
 from datetime import timedelta
 
 SIMPLE_JWT = {
-    'ACCESS_TOKEN_LIFETIME': timedelta(hours=1),
+    'ACCESS_TOKEN_LIFETIME': timedelta(seconds=10),
     'REFRESH_TOKEN_LIFETIME': timedelta(weeks=2),
 }
