+++ final-pjt-front/paradiso_vue/src/store/index.js
@@ -1,6 +1,7 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 import createPersistedState from 'vuex-persistedstate'
+import axios from 'axios'
 
 Vue.use(Vuex)
 
@@ -38,6 +39,23 @@ export default new Vuex.Store({
 
     select_view(context, bool) {
       context.commit("SELECT_VIEW", bool)
+    },
+
+    refresh() {
+      const refreshToken = localStorage.getItem("refreshToken")
+      axios({
+        method: "post",
+        url: "http://127.0.0.1:8000/api/token/refresh/",
+        data: {'refresh': refreshToken},
+      })
+        .then((res) => {
+          localStorage.setItem("accessToken", res.data.access)
+        })
+        .catch(() => {
+          localStorage.removeItem("accessToken")
+          localStorage.removeItem("refreshToken")
+          this.$router.push({ name: "LogInView" })
+        })
     }
   },
   modules: {
